+++ Assets/Scripts/Model/Game/Component/StepC.cs
@@ -2,7 +2,7 @@
 {
     public struct StepsC
     {
-        public float Steps { get; internal set; }
+        public double Steps { get; internal set; }
 
         public bool HaveAnySteps => Steps > 0;
     }
